make VGGLayers class and remove vgg19.py
