+++ chainer/links/__init__.py
@@ -51,7 +51,7 @@ from chainer.links.model.vision.resnet import ResNet101Layers  # NOQA
 from chainer.links.model.vision.resnet import ResNet152Layers  # NOQA
 from chainer.links.model.vision.resnet import ResNet50Layers  # NOQA
 from chainer.links.model.vision.vgg import VGG16Layers  # NOQA
-from chainer.links.model.vision.vgg19 import VGG19Layers  # NOQA
+from chainer.links.model.vision.vgg import VGG19Layers  # NOQA
 from chainer.links.normalization.batch_normalization import BatchNormalization  # NOQA
 from chainer.links.normalization.batch_renormalization import BatchRenormalization  # NOQA
 from chainer.links.normalization.group_normalization import GroupNormalization  # NOQA
